Move user identification heuristics to `lib.user.member_request`
